backend/requirements.txt
Pillow
fastapi
uvicorn
transformers
python-multipart
tensorflow
easyocr
keras_ocr
face_recognition